clean up the pipeline
