+++ scripts/project/test-automatic-submission/config.js
@@ -1,0 +1,50 @@
+// All configuration constants. Single source of truth.
+export const ORG_UNIT =
+  "http://data.lblod.info/id/bestuurseenheden/be278471a2a318edba32e7ac4294c0eafbe4c8077a34dcbb9c2e43211d4a78a6";
+export const ORGAN_ABSTRACT =
+  "http://data.lblod.info/id/bestuursorganen/06c2b56ed7b49d146337f6db044204f19c34c4242deb3b4e142dbf925d733eda";
+export const ORGAN_IN_TIJD =
+  "http://data.lblod.info/id/bestuursorganen/86af6b8e417005a56d3c2437f6345c4676dfeee1d7162e4f1a2899240b6d0476";
+export const ORGAN_LABEL = "Gemeenteraad Mechelen";
+
+export const BESLUITENLIJST_TYPE =
+  "https://data.vlaanderen.be/id/concept/BesluitDocumentType/3fa67785-ffdc-4b30-8880-2b99d97b4dee";
+
+export const SPARQL_ENDPOINT = "http://virtuoso:8890/sparql";
+export const MELDING_ENDPOINT = "http://identifier/melding";
+
+export const STATUS_CONCEPT =
+  "http://lblod.data.gift/concepts/79a52da4-f491-4e2f-9374-89a13cde8ecd";
+export const STATUS_INZENDBAAR =
+  "http://lblod.data.gift/concepts/f6330856-e261-430f-b949-8e510d20d0ff";
+export const STATUS_VERSTUURD =
+  "http://lblod.data.gift/concepts/9bd8d86d-bb10-4456-a84e-91e9507c374c";
+
+export const JOB_SCHEDULED = "http://redpencil.data.gift/id/concept/JobStatus/scheduled";
+export const JOB_BUSY = "http://redpencil.data.gift/id/concept/JobStatus/busy";
+export const JOB_SUCCESS = "http://redpencil.data.gift/id/concept/JobStatus/success";
+export const JOB_FAILED = "http://redpencil.data.gift/id/concept/JobStatus/failed";
+
+export const DL_READY = "http://lblod.data.gift/file-download-statuses/ready-to-be-cached";
+export const DL_ONGOING = "http://lblod.data.gift/file-download-statuses/ongoing";
+export const DL_SUCCESS = "http://lblod.data.gift/file-download-statuses/success";
+export const DL_FAILURE = "http://lblod.data.gift/file-download-statuses/failure";
+
+export const JOB_OPERATION =
+  "http://lblod.data.gift/id/jobs/concept/JobOperation/automaticSubmissionFlow";
+export const TASK_OPS = {
+  register: "http://lblod.data.gift/id/jobs/concept/TaskOperation/register",
+  download: "http://lblod.data.gift/id/jobs/concept/TaskOperation/download",
+  import: "http://lblod.data.gift/id/jobs/concept/TaskOperation/import",
+  enrich: "http://lblod.data.gift/id/jobs/concept/TaskOperation/enrich",
+  validate: "http://lblod.data.gift/id/jobs/concept/TaskOperation/validate",
+  "form-data-generate":
+    "http://lblod.data.gift/id/jobs/concept/TaskOperation/form-data-generate",
+};
+
+export const DOC_URI_BASE = "http://test.local/besluitenlijsten/";
+export const PAGE_PORT = 8888;
+export const POLL_INTERVAL = 2000;
+export const POLL_TIMEOUT = 150000;
+
+export const RESULT_FILE = "/data/app/test-automatic-submission/last-run.json";
